+++ src/shared/Logging/ExtendedDebugLogLevel.cs
@@ -1,0 +1,7 @@
+namespace Cmdb2Monitoring.Logging;
+
+public enum ExtendedDebugLogLevel
+{
+    Basic = 0,
+    Verbose = 1
+}